rename folder: `src/instrumentation` -> `src/registerSQLFunctions`
