+++ app/src/app/regionen/[regionSlug]/_mapData/mapDataSources/sources.const.ts
@@ -1,5 +1,5 @@
 import { getTilesUrl } from '@/src/app/_components/utils/getTilesUrl'
-import { SIMPLIFY_MAX_ZOOM } from '@/src/instrumentation/registerGeneralizationFunctions'
+import { SIMPLIFY_MAX_ZOOM } from '@/src/registerSQLFunctions/registerGeneralizationFunctions'
 import { MapDataSource } from '../types'
 import { apiKeyMapbox, apiKeyMapillary } from './apiKeys.const'
 import { SourceExportApiIdentifier } from './export/exportIdentifier'
