+++ app/src/registerSQLFunctions/registerGeneralizationFunctions.ts
