+++ app/src/registerSQLFunctions/registerSQLFunctions.ts
